LCU Protocol/Summoner.cs fix.

+++ Sources/LeagueBot/LeagueBot/LCU/Protocol/Summoner.cs
@@ -24,7 +24,7 @@ namespace LeagueBot.LCU.Protocol
 
         public RerollPoints RerollPoints;
 
-        public int summonerId;
+        public long summonerId;
 
         public short summonerLevel;
